requirements.txt
# Core
langchain>=0.1.0
langchain-community
langchain-groq
langchain-huggingface
langchain-core
sentence-transformer

# Vector Store
faiss-cpu

# Document Processing
pypdf

# CLI
typer>=0.9.0
rich>=13.0.0

# Web Interface
gradio>=4.0.0

# Utils
python-dotenv

